+++ src/gql/interfaces/elements/PodcastInterface.php
@@ -5,6 +5,8 @@ namespace vnali\studio\gql\interfaces\elements;
 use Craft;
 use craft\gql\GqlEntityRegistry;
 use craft\gql\interfaces\Element;
+use craft\gql\interfaces\elements\User;
+use craft\helpers\Gql;
 use GraphQL\Type\Definition\InterfaceType;
 use GraphQL\Type\Definition\Type;
 use vnali\studio\elements\Podcast;
@@ -56,7 +58,7 @@ class PodcastInterface extends Element
      */
     public static function getFieldDefinitions(): array
     {
-        return Craft::$app->getGql()->prepareFieldDefinitions(array_merge(parent::getFieldDefinitions(), [
+        return Craft::$app->getGql()->prepareFieldDefinitions(array_merge(parent::getFieldDefinitions(), self::getConditionalFields(),[
             'copyright' => [
                 'name' => 'copyright',
                 'type' => Type::string(),
@@ -128,5 +130,29 @@ class PodcastInterface extends Element
                 'description' => 'GUID',
             ],
         ]), self::getName());
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected static function getConditionalFields(): array
+    {
+        $fields = [];
+        if (Gql::canQueryUsers()) {
+            $fields = array_merge($fields, [
+                'uploaderId' => [
+                    'name' => 'uploaderId',
+                    'type' => Type::int(),
+                    'description' => 'The ID of the uploader of this podcast.',
+                ],
+                'uploader' => [
+                    'name' => 'uploader',
+                    'type' => User::getType(),
+                    'description' => 'The podcast’s uploader.',
+                    'complexity' => Gql::eagerLoadComplexity(),
+                ],
+            ]);
+        }
+        return $fields;
     }
 }